--- src/main/java/org/prgms/voucher/SpringVoucherApplication.java
@@ -1,14 +1,0 @@
-package org.prgms.voucher;
-
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
-
-@SpringBootApplication
-public class SpringVoucherApplication {
-
-    public static void main(String[] args) {
-        ApplicationContext applicationContext = new AnnotationConfigApplicationContext(SpringVoucherApplication.class);
-        applicationContext.getBean(VoucherProgram.class).run();
-    }
-}
