--- src/main/java/org/prgms/voucher/entity/Voucher.java
@@ -1,10 +1,0 @@
-package org.prgms.voucher.entity;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-public interface Voucher extends Serializable {
-    long discount(long beforeDiscount);
-
-    UUID getVoucherId();
-}
