--- src/main/java/org/prgms/voucher/exception/WrongDiscountAmountException.java
@@ -1,7 +1,0 @@
-package org.prgms.voucher.exception;
-
-public class WrongDiscountAmountException extends Exception {
-    public WrongDiscountAmountException() {
-        super("[ERROR] 올바른 할인금액이 아닙니다.");
-    }
-}
