--- src/main/java/org/prgms/voucher/exception/WrongDiscountPercentException.java
@@ -1,7 +1,0 @@
-package org.prgms.voucher.exception;
-
-public class WrongDiscountPercentException extends Exception {
-    public WrongDiscountPercentException() {
-        super("[ERROR] 올바른 할인퍼센트가 아닙니다.");
-    }
-}
