--- src/main/java/org/prgms/voucher/exception/WrongInputMenuException.java
@@ -1,7 +1,0 @@
-package org.prgms.voucher.exception;
-
-public class WrongInputMenuException extends Exception {
-    public WrongInputMenuException() {
-        super("[ERROR] 올바른 메뉴 입력이 아닙니다.");
-    }
-}
