--- src/main/java/org/prgms/voucher/exception/WrongInputVoucherCommandException.java
@@ -1,7 +1,0 @@
-package org.prgms.voucher.exception;
-
-public class WrongInputVoucherCommandException extends Exception {
-    public WrongInputVoucherCommandException() {
-        super("[ERROR] 올바른 바우처 커맨드가 아닙니다.");
-    }
-}
