--- src/main/java/org/prgms/voucher/repository/MemoryVoucherRepository.java
@@ -1,27 +1,0 @@
-package org.prgms.voucher.repository;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-
-import org.prgms.voucher.entity.Voucher;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class MemoryVoucherRepository implements VoucherRepository {
-
-    private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
-
-    @Override
-    public Voucher save(Voucher voucher) {
-        storage.put(voucher.getVoucherId(), voucher);
-        return voucher;
-    }
-
-    @Override
-    public List<Voucher> findAll() {
-        return new ArrayList<>(storage.values());
-    }
-}
