--- src/main/java/org/prgms/voucher/repository/VoucherRepository.java
@@ -1,11 +1,0 @@
-package org.prgms.voucher.repository;
-
-import java.util.List;
-
-import org.prgms.voucher.entity.Voucher;
-
-public interface VoucherRepository {
-    Voucher save(Voucher voucher);
-
-    List<Voucher> findAll();
-}
