--- src/main/java/org/prgms/voucher/view/InputView.java
@@ -1,12 +1,0 @@
-package org.prgms.voucher.view;
-
-import org.prgms.voucher.entity.VoucherType;
-
-public interface InputView {
-
-    String inputMenu();
-
-    String inputVoucherCommand();
-
-    long inputDiscountValue(VoucherType voucherType);
-}
