+++ src/main/java/org/prgms/voucherProgram/SpringVoucherApplication.java
@@ -1,0 +1,14 @@
+package org.prgms.voucherProgram;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+
+@SpringBootApplication
+public class SpringVoucherApplication {
+
+    public static void main(String[] args) {
+        ApplicationContext applicationContext = new AnnotationConfigApplicationContext(SpringVoucherApplication.class);
+        applicationContext.getBean(VoucherProgram.class).run();
+    }
+}
