+++ src/main/java/org/prgms/voucherProgram/VoucherProgram.java
@@ -1,0 +1,74 @@
+package org.prgms.voucherProgram;
+
+import org.prgms.voucherProgram.entity.MenuType;
+import org.prgms.voucherProgram.entity.voucher.Voucher;
+import org.prgms.voucherProgram.entity.voucher.VoucherType;
+import org.prgms.voucherProgram.service.VoucherService;
+import org.prgms.voucherProgram.view.Console;
+import org.prgms.voucherProgram.view.InputView;
+import org.prgms.voucherProgram.view.OutputView;
+import org.springframework.stereotype.Component;
+
+@Component
+public class VoucherProgram {
+    private final VoucherService voucherService;
+    private final InputView inputView;
+    private final OutputView outputView;
+
+    public VoucherProgram(VoucherService voucherService, Console console) {
+        this.voucherService = voucherService;
+        this.inputView = console;
+        this.outputView = console;
+    }
+
+    public void run() {
+        boolean isNotEndProgram = true;
+
+        while (isNotEndProgram) {
+            MenuType menuType = inputMenu();
+            switch (menuType) {
+                case EXIT:
+                    isNotEndProgram = false;
+                    break;
+                case LIST:
+                    outputView.printAllVoucher(voucherService.findAllVoucher());
+                    break;
+                case CREATE:
+                    VoucherType voucherType = inputVoucherCommand();
+                    Voucher voucher = createVoucher(voucherType);
+                    outputView.printVoucher(voucher);
+            }
+        }
+    }
+
+    private Voucher createVoucher(VoucherType voucherType) {
+        while (true) {
+            try {
+                long discountValue = inputView.inputDiscountValue(voucherType);
+                return voucherService.create(voucherType, discountValue);
+            } catch (Exception e) {
+                outputView.printError(e.getMessage());
+            }
+        }
+    }
+
+    private VoucherType inputVoucherCommand() {
+        while (true) {
+            try {
+                return VoucherType.findByCommand(inputView.inputVoucherCommand());
+            } catch (Exception e) {
+                outputView.printError(e.getMessage());
+            }
+        }
+    }
+
+    private MenuType inputMenu() {
+        while (true) {
+            try {
+                return MenuType.of(inputView.inputMenu());
+            } catch (Exception e) {
+                outputView.printError(e.getMessage());
+            }
+        }
+    }
+}
