+++ src/main/java/org/prgms/voucherProgram/entity/voucher/Voucher.java
@@ -1,0 +1,10 @@
+package org.prgms.voucherProgram.entity.voucher;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+public interface Voucher extends Serializable {
+    long discount(long beforeDiscount);
+
+    UUID getVoucherId();
+}
