+++ src/main/java/org/prgms/voucherProgram/exception/WrongDiscountAmountException.java
@@ -1,0 +1,7 @@
+package org.prgms.voucherProgram.exception;
+
+public class WrongDiscountAmountException extends Exception {
+    public WrongDiscountAmountException() {
+        super("[ERROR] 올바른 할인금액이 아닙니다.");
+    }
+}
