+++ src/main/java/org/prgms/voucherProgram/exception/WrongDiscountPercentException.java
@@ -1,0 +1,7 @@
+package org.prgms.voucherProgram.exception;
+
+public class WrongDiscountPercentException extends Exception {
+    public WrongDiscountPercentException() {
+        super("[ERROR] 올바른 할인퍼센트가 아닙니다.");
+    }
+}
