+++ src/main/java/org/prgms/voucherProgram/repository/voucher/MemoryVoucherRepository.java
@@ -1,0 +1,27 @@
+package org.prgms.voucherProgram.repository.voucher;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+
+import org.prgms.voucherProgram.entity.voucher.Voucher;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class MemoryVoucherRepository implements VoucherRepository {
+
+    private final Map<UUID, Voucher> storage = new ConcurrentHashMap<>();
+
+    @Override
+    public Voucher save(Voucher voucher) {
+        storage.put(voucher.getVoucherId(), voucher);
+        return voucher;
+    }
+
+    @Override
+    public List<Voucher> findAll() {
+        return new ArrayList<>(storage.values());
+    }
+}
