--- src/test/java/org/prgms/voucher/SpringVoucherApplicationTests.java
@@ -1,13 +1,0 @@
-package org.prgms.voucher;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SpringVoucherApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
