+++ src/test/java/org/prgms/voucherProgram/SpringVoucherApplicationTests.java
@@ -1,0 +1,13 @@
+package org.prgms.voucherProgram;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SpringVoucherApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
